fix: sometimes keys would stop responding
In case of a collision, for example due to automatically going down, the
entire previous state of the player was restored, including the timer
which determines whether the player should go down next frame. As a
result, the player tried to go down the next frame as well. The player
ignores player input on the frames in which it goes down, that's
supposed to be very rarely, but now it was every frame.