port auto_rebuild
To look at the elm files in the src folder, which is where elm now
recommends to put source files, and to use fswatcher instead of fswatch.
I just like fswatcher better for the irrational reason that it is
written in Haskell.

I am using this instead of "elm reactor" so that I can use my custom
html.