fix gravity
Player.Msg used to be a product containing both a time delta and the key
pressed if any. Switching away from signals to subscriptions, it is now
easier to use a sum type instead. The code had two functions,
time_passed and keys_are_pressed, which seemed to handle each type of
event, so I only sent time events to time_passed and key events to
keys_are_pressed. However, it turns out that the gravity was processed
in keys_are_pressed, not time_passed, and thus gravity did not apply
unless the player pressed a key which the game didn't use, such as
spacebar.