fix repeated rotation/shape-shifting
Pressing either of those keys several times in a row would have no
effect until the next gravity tick or movement. This is keeping either
of those buttons held down in order to constantly rotate or change shape
looked weird, so there was some code to detect and prevent that, the
last_keys code. But since the new subscription code was not receiving
key up events like the signals code was, pressing, releasing, and
pressing the key again looked the same as pressing, holding, and
receiving the system's key-repeat events.